+++ src/panoptes/pocs/state/machine.py
@@ -10,6 +10,8 @@ from panoptes.utils.library import load_module
 from panoptes.utils.serializers import from_yaml
 from panoptes.utils.utils import listify
 
+from panoptes.pocs.utils.telemetry import StateMachineState
+
 
 class PanStateMachine(Machine):
     """A finite state machine for PANOPTES.
@@ -209,9 +211,15 @@ class PanStateMachine(Machine):
         state_changed = transition_method()
         if state_changed:
             self.logger.success(f"Finished with {self.state} state")
-            self.db.insert_current(
-                "state", {"source": self.state, "dest": self.next_state}, store_permanently=False
-            )
+            data = {"source": self.state, "dest": self.next_state}
+
+            # Record to telemetry server.
+            try:
+                state_reading = StateMachineState(**data)
+                if hasattr(self, "record_telemetry"):
+                    self.record_telemetry(state_reading)
+            except Exception as e:
+                self.logger.warning(f"Could not record state telemetry: {e!r}")
 
         return state_changed