+++ tests/test_base.py
@@ -1,15 +1,8 @@
-from panoptes.utils.database import PanDB
-
 from panoptes.pocs.base import PAN_CONFIG_CACHE, PanBase
 
 
 def test_with_logger():
     PanBase()
-
-
-def test_with_db():
-    base = PanBase(db=PanDB(db_type="memory", db_name="tester"))
-    assert isinstance(base, PanBase)
 
 
 def test_remember_config():